move birthday check into own file
